Refactor logic.
